requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
gunicorn==20.1.0
Markdown==3.3.4
Pillow==8.4.0
psycopg2==2.9.1
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
